// src/types/carpetas.ts
export type Juzgado = {
  id: string;
  tipo: string;
  ciudad: string;
  url: string;
};